Clipping HWRT reflection near-field rays by a culling sphere determined by ray-tracing culling radius. Fixes specular occlusion leaking when transiiton between near and far-field.
#rb krzysztof.narkowicz, daniel.wright
#preflight 6160460f4295c40001ce3b5e
#lockdown michal.valient

#ROBOMERGE-AUTHOR: patrick.kelly
#ROBOMERGE-SOURCE: CL 17760821 via CL 17985606 via CL 18368064 via CL 18368119
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Staging -> Release-Engine-Test) (v895-18170469)

[CL 18368146 by patrick kelly in ue5-release-engine-test branch]

+++ Engine/Source/Runtime/Renderer/Private/Lumen/LumenReflections.h
@@ -27,7 +27,7 @@ BEGIN_SHADER_PARAMETER_STRUCT(FLumenReflectionTracingParameters, )
 	SHADER_PARAMETER(float, ReflectionSmoothBias)
 
 	SHADER_PARAMETER_RDG_TEXTURE(Texture2D<float4>, RayBuffer)
-	SHADER_PARAMETER_RDG_TEXTURE(Texture2D<float>, RayTraceDistance)
+	SHADER_PARAMETER_RDG_TEXTURE(Texture2D<uint>, RayTraceDistance)
 	SHADER_PARAMETER_RDG_TEXTURE(Texture2D, DownsampledDepth)
 
 	SHADER_PARAMETER_RDG_TEXTURE(Texture2D, TraceHit)
